requirements.txt
pydantic>=2.8.2
pandas
requests
logfire>=0.51.0
openai==1.42.0

python-dotenv
